requirements.txt
accelerate~=1.7.0
diffusers~=0.33.1
geopandas~=1.1.1
huggingface-hub~=0.34.4
matplotlib~=3.10.3
natsort==8.4.0
numpy~=2.2.6
opencv-python==4.12.0.88
osmnx==2.0.5
packaging==25.0
pillow~=11.2.1
pyproj~=3.7.1
pyrender==0.1.45
requests~=2.32.3
shapely~=2.1.1
torch~=2.7.0
tqdm==4.67.1
transformers~=4.55.0
trimesh==4.7.1

datasets~=3.6.0
scikit-image~=0.25.2
pygame~=2.6.1
dotenv~=0.9.9
python-dotenv~=1.1.0
controlnet_aux~=0.0.10
torchvision~=0.22.0
ultralytics~=8.3.173